norminette

+++ ft_strlcpy.c
@@ -1,17 +1,29 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlcpy.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: skawanis <skawanis@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/01/09 05:31:41 by skawanis          #+#    #+#             */
+/*   Updated: 2023/01/09 05:31:58 by skawanis         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "mylib.h"
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-        const unsigned int      srclen = ft_strlen(src);
+	const size_t	srclen = ft_strlen(src);
 
-        if (srclen < size)
-        {
-                ft_memcpy(dest, src, srclen + 1);
-        }
-        else if (size != 0)
-        {
-                ft_memcpy(dest, src, size - 1);
-                dest[size - 1] = '\0';
-        }
-        return (ft_strlen(src));
+	if (srclen < size)
+	{
+		ft_memcpy(dest, src, srclen + 1);
+	}
+	else if (size != 0)
+	{
+		ft_memcpy(dest, src, size - 1);
+		dest[size - 1] = '\0';
+	}
+	return (ft_strlen(src));
 }
